# Rework HashTable implementation