--- a/nanoFramework.System.Collections/Properties/AssemblyInfo.cs
+++ b/nanoFramework.System.Collections/Properties/AssemblyInfo.cs
@@ -12,7 +12,7 @@ using System.Runtime.InteropServices;
 
 ////////////////////////////////////////////////////////////////
 // update this whenever the native assembly signature changes //
-[assembly: AssemblyNativeVersion("100.0.0.1")]
+[assembly: AssemblyNativeVersion("100.0.1.0")]
 ////////////////////////////////////////////////////////////////
 
 // Setting ComVisible to false makes the types in this assembly not visible