--- a/nanoFramework.System.Collections/nanoFramework.System.Collections.nfproj
+++ b/nanoFramework.System.Collections/nanoFramework.System.Collections.nfproj
@@ -37,9 +37,17 @@
     <NFMDP_PE_ExcludeClassByName Include="ThisAssembly">
       <InProject>false</InProject>
     </NFMDP_PE_ExcludeClassByName>
+    <NFMDP_PE_ExcludeClassByName Include="DictionaryEntry">
+      <InProject>false</InProject>
+    </NFMDP_PE_ExcludeClassByName>
+    <NFMDP_PE_ExcludeClassByName Include="IDictionaryEnumerator">
+      <InProject>false</InProject>
+    </NFMDP_PE_ExcludeClassByName>
   </ItemGroup>
   <Import Project="$(NanoFrameworkProjectSystemPath)NFProjectSystem.props" Condition="Exists('$(NanoFrameworkProjectSystemPath)NFProjectSystem.props')" />
   <ItemGroup>
+    <Compile Include="Collections\Hashtable.Bucket.cs" />
+    <Compile Include="Collections\IDictionaryEnumerator.cs" />
     <Compile Include="Collections\DictionaryEntry.cs" />
     <Compile Include="Collections\Hashtable.cs" />
     <Compile Include="Collections\IDictionary.cs" />