… Tests/HashtableTests/HashtableTests.cs
… nanoFramework.System.Collections/Collections/Hashtable.cs