## Description
- Constructors are now aligned with .NET as all other methods.
- Remove unnecessary properties related with load factor.
- Add parameter checks that were missing in most methods.
- Improve several methods and constructors.
- Add IDictionaryEnumerator to follow .NET.
- Move hot path methods to native.
- Improve comments for Intellisense.
- Add more Unit Tests.

## Motivation and Context
- Need to align with .NET and more important improve performance.

## How Has This Been Tested?<!-- (IF APPLICABLE) -->
- Add more unit tests.

Execution comparison running Unit Tests in an ESP32 WROVER KIT
Ctor_IDictionary: 3040ms -> 486ms
Ctor_Int: 640ms -> 460ms
Add_ReferenceType: 20ms -> 8ms
Add_ClearRepeatedly: 40ms -> 27ms
AddRemove_LargeAmountNumbers: 12460ms -> 2400ms
Clone: 2170ms -> 1000ms
Clone_IsShallowCopy: 150ms -> 46ms
Clone_HashtableCastedToInterfaces: 840ms -> 345ms
ContainsKey: 1750ms -> 731ms
ContainsKey_EqualObjects: 20ms -> 16ms
Keys_ModifyingHashtable_ModifiesCollection: 1010ms -> 474ms
Values_ModifyingHashtable_ModifiesCollection: 1020ms -> 480ms
Hashtable_Add_Contains_Remove_Test1: 160ms -> 58ms
Hashtable_Add_Clear: 130ms -> 68ms
Hashtable_CheckKeys: 150ms -> 40ms
Hashtable_CheckValues: 260ms -> 134ms
Hashtable_Count: 150ms -> 51ms
Hashtable_Duplicate: 270ms -> 66ms
Hashtable_CopyTo: 390ms -> 115ms

## Screenshots<!-- (IF APPROPRIATE): -->

## Types of changes
<!--- What types of changes does this PR introduce? Put an `x` in all the boxes that apply: -->
- [x] Improvement (non-breaking change that improves a feature, code or algorithm)
- [ ] Bug fix (non-breaking change which fixes an issue with code or algorithm)
- [ ] New feature (non-breaking change which adds functionality to code)
- [x] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Config and build (change in the configuration and build system, has no impact on code or features)
- [ ] Dependencies (update dependencies and changes associated, has no impact on code or features)
- [x] Unit Tests (add new Unit Test(s) or improved existing one(s), has no impact on code or features)
- [ ] Documentation (changes or updates in the documentation, has no impact on code or features)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
<!--- PLEASE PLEASE PLEASE don't tick all of them just because -->
- [x] My code follows the code style of this project (only if there are changes in source code).
- [x] My changes require an update to the documentation (there are changes that require the docs website to be updated).
- [ ] I have updated the documentation accordingly (the changes require an update on the docs in this repo).
- [x] I have read the [CONTRIBUTING](https://github.com/nanoframework/.github/blob/main/CONTRIBUTING.md) document.
- [x] I have tested everything locally and all new and existing tests passed (only if there are changes in source code).
- [x] I have added new tests to cover my changes.