@josesimoes 2022-03-08
Unit Tests are failing because WIN32 nanoCLR version is behind this.